+++ static_precompiler/templatetags/base.py
@@ -1,6 +1,7 @@
 from django.core.cache import cache
 from django.template import Node
-from static_precompiler.settings import USE_CACHE, CACHE_TIMEOUT
+from django.templatetags.static import static
+from static_precompiler.settings import USE_CACHE, CACHE_TIMEOUT, PREPEND_STATIC_URL
 from static_precompiler.utils import get_cache_key, get_hexdigest
 
 
@@ -15,7 +16,10 @@ class BaseInlineNode(Node):
         source = self.nodelist.render(context)
 
         if USE_CACHE:
-            cache_key = get_cache_key(get_hexdigest(source))
+            cache_key = get_cache_key("{0}.{1}".format(
+                self.__class__.__name__,
+                get_hexdigest(source)
+            ))
             cached = cache.get(cache_key, None)
             if cached is not None:
                 return cached