+++ static_precompiler/templatetags/sass.py
@@ -1,7 +1,7 @@
 from django.template.base import Library
 from static_precompiler.compilers import SASS
 from static_precompiler.templatetags.base import BaseInlineNode
-from static_precompiler.utils import prepend_static_url
+from static_precompiler.templatetags.compile_static import compile_tag
 
 
 register = Library()
@@ -15,13 +15,12 @@ class InlineSASSNode(BaseInlineNode):
 
 #noinspection PyUnusedLocal
 @register.tag(name="inlinesass")
-def do_inlinecoffeescript(parser, token):
+def do_inlinesass(parser, token):
     nodelist = parser.parse(("endinlinesass",))
     parser.delete_first_token()
     return InlineSASSNode(nodelist)
 
 
-@register.simple_tag
-def sass(path):
-    return prepend_static_url(compiler.compile(str(path)))
-
+@register.simple_tag(name="sass")
+def sass_tag(source_path):
+    return compile_tag(source_path, compiler)