+++ static_precompiler/tests/__init__.py
@@ -7,6 +7,7 @@ def suite():
     from static_precompiler.tests import test_less
     from static_precompiler.tests import test_coffeescript
     from static_precompiler.tests import test_scss
+    from static_precompiler.tests import test_templatetags
     from static_precompiler.tests import test_static_precompiler_watch
 
     test_suite = unittest.TestSuite()
@@ -15,6 +16,7 @@ def suite():
     test_suite.addTests(test_less.suite())
     test_suite.addTests(test_coffeescript.suite())
     test_suite.addTests(test_scss.suite())
+    test_suite.addTests(test_templatetags.suite())
     test_suite.addTests(test_static_precompiler_watch.suite())
     return test_suite